refactor: dedupe job config datasets and mpp verifier harness
- Share the dataset matrix in config/datasets.yaml; run-benchmark.ts injects
  it into any job config without its own datasets block (job.local.oracle.dev
  and the production template keep explicit overrides).
- Move the byte-identical MPP verifier harness (Dockerfile, tsconfig, test.sh,
  verify.sh, quality check, verifier_utils, client harness) into shared/mpp/,
  synced into each tasks/mpp/<task>/ by sync-shared.
- Split each task's client.py into a shared client_lib.py harness plus a small
  per-task scenario defining run_task; scenarios are byte-equivalent to the
  previous per-task run_task bodies.
- Convert scripts/sync-shared.mjs to scripts/sync-shared.ts so tsc checks it.
- Add MPP dataset freshness check to CI.
