+++ .github/workflows/ci.yml
@@ -42,3 +42,6 @@ jobs:
 
       - name: Verify dataset freshness
         run: npm run check:dataset
+
+      - name: Verify MPP dataset freshness
+        run: npm run check:dataset -- --tasks tasks/mpp
