+++ config/job.local.agent.dev.yaml
@@ -20,12 +20,4 @@ agents:
     env:
       ANTHROPIC_API_KEY: ${ANTHROPIC_API_KEY:-}
       ANTHROPIC_AUTH_TOKEN: ${ANTHROPIC_AUTH_TOKEN:-}
-datasets:
-  - path: tasks/tempo
-    task_names:
-      - "*-base"
-      - "*-docs"
-      - "*-mcp"
-  - path: tasks/mpp
-    task_names:
-      - "server-*"
+# Datasets are injected from config/datasets.yaml; declare `datasets:` here to override.
