+++ config/job.local.oracle.yaml
@@ -16,12 +16,4 @@ verifier:
     REWARDKIT_JUDGE: ${REWARDKIT_JUDGE:-anthropic/claude-haiku-4-5}
 agents:
   - name: oracle
-datasets:
-  - path: tasks/tempo
-    task_names:
-      - "*-base"
-      - "*-docs"
-      - "*-mcp"
-  - path: tasks/mpp
-    task_names:
-      - "server-*"
+# Datasets are injected from config/datasets.yaml; declare `datasets:` here to override.
