+++ scripts/sync-shared.ts
@@ -5,6 +5,21 @@ import { fileURLToPath } from "node:url";
 
 const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 const tasksDir = path.join(root, "tasks", "tempo");
+const mppTasksDir = path.join(root, "tasks", "mpp");
+const mppSharedDir = path.join(root, "shared", "mpp");
+// Shared MPP verifier harness files copied into every tasks/mpp/<task>/.
+// Task-specific files (task.toml, instruction.md, solution sources,
+// tests/support/client.py, tests/correctness/criteria.py,
+// tests/quality/reward.toml) stay in the task directory.
+const mppSharedFiles = [
+  { source: "environment/Dockerfile", executable: false },
+  { source: "solution/tsconfig.json", executable: false },
+  { source: "tests/test.sh", executable: true },
+  { source: "tests/correctness/verify.sh", executable: true },
+  { source: "tests/quality/check.py", executable: false },
+  { source: "tests/support/client_lib.py", executable: false },
+  { source: "tests/support/verifier_utils.py", executable: false },
+];
 const tempoDocsUrl = "http://tempo-docs:3000/developers";
 const tempoMcpUrl = "https://mcp.tempo.xyz";
 const defaultTurnCutoffs = "20=1.0,40=0.8,60=0.5,80=0.2,*=0.0";
@@ -374,7 +389,7 @@ digest = ${JSON.stringify(digests.get(name) ?? placeholderDigest)}
   writeFile(
     path.join(tasksDir, "dataset.toml"),
     `# Dataset manifest for tempo/tempo-bench-v1
-# Generated by scripts/sync-shared.mjs. Run \`harbor sync tasks/tempo\` to refresh digests.
+# Generated by scripts/sync-shared.ts. Run \`harbor sync tasks/tempo\` to refresh digests.
 
 [dataset]
 name = "tempo/tempo-bench-v1"
@@ -492,8 +507,8 @@ function linkFile(source, destination) {
   fs.symlinkSync(relativeSymlinkTarget(destination, source), destination, "file");
 }
 
-function readStringTable(filePath, tableName) {
-  const values = {};
+function readStringTable(filePath: string, tableName: string): Record<string, string> {
+  const values: Record<string, string> = {};
   let currentTable = "";
 
   for (const line of fs.readFileSync(filePath, "utf8").split(/\r?\n/)) {
@@ -592,6 +607,27 @@ function taskDirs() {
     .filter((dir) => fs.existsSync(path.join(dir, "task.toml")));
 }
 
+function mppTaskDirs() {
+  if (!fs.existsSync(mppTasksDir)) return [];
+  return fs
+    .readdirSync(mppTasksDir, { withFileTypes: true })
+    .filter((entry) => entry.isDirectory())
+    .map((entry) => path.join(mppTasksDir, entry.name))
+    .filter((dir) => fs.existsSync(path.join(dir, "task.toml")));
+}
+
+function syncMppTasks() {
+  const tasks = mppTaskDirs();
+  for (const taskDir of tasks) {
+    for (const file of mppSharedFiles) {
+      const destination = path.join(taskDir, file.source);
+      copyFile(path.join(mppSharedDir, file.source), destination);
+      if (file.executable) fs.chmodSync(destination, 0o755);
+    }
+  }
+  return tasks.length;
+}
+
 materializeTaskMatrix();
 writeDatasetManifest();
 const tasks = taskDirs();
@@ -650,6 +686,8 @@ for (const taskDir of tasks) {
   assertComposeBuildContexts(taskDir);
 }
 
+const mppTaskCount = syncMppTasks();
+
 console.log(
-  `Synced verifier and linked localnet assets into ${tasks.length} task(s).`,
+  `Synced verifier and linked localnet assets into ${tasks.length} task(s); synced MPP harness into ${mppTaskCount} task(s).`,
 );
