+++ tasks/tempo/dataset.toml
@@ -1,5 +1,5 @@
 # Dataset manifest for tempo/tempo-bench-v1
-# Generated by scripts/sync-shared.mjs. Run `harbor sync tasks/tempo` to refresh digests.
+# Generated by scripts/sync-shared.ts. Run `harbor sync tasks/tempo` to refresh digests.
 
 [dataset]
 name = "tempo/tempo-bench-v1"
